refactor: 未使用機能を削除
- ViewAngleVisualizer を機能ごと削除（ファイル・設定・キーバインド・イベント）
  - ViewAngleVisualizer.ts 削除
  - ViewAngleVisualizationConfig 削除
  - GameEventBus から VIEW_ANGLE_TOGGLED/VIS_TOGGLE_VIEW_ANGLE 削除
  - GameConstants から TOGGLE_VIEW_ANGLE キー定数削除
  - MobileControls の視野角ボタン削除
  - GameController/InputHandler/VisualizationSync の関連コード削除
- GameEventBus のデバッグ機能を削除（eventHistory・6メソッド）
- ObstacleExporter.ts を削除し ObstacleData 型を MapGenerator.ts に統合
