--- src/model/ObstacleExporter.ts
@@ -1,11 +1,0 @@
-import type { LineSegment } from './LineSegment';
-
-/**
- * Obstacle data structure (now stores LineSegments directly)
- */
-export interface ObstacleData {
-  id: number;
-  segments: LineSegment[];
-}
-
-
